// types.ts
export enum ModuleType {
  DASHBOARD = 'DASHBOARD',
  OPERATIONS = 'OPERATIONS',
  SALES = 'SALES',
  RETAIL = 'RETAIL',
  CLIENTS = 'CLIENTS',
  CRM = 'CRM',
  SUPPLIERS = 'SUPPLIERS',
  EVENTS = 'EVENTS',
  INVENTORY = 'INVENTORY',
  PURCHASES = 'PURCHASES',
  FINANCE = 'FINANCE',
  HR = 'HR',
  ACCOUNTING = 'ACCOUNTING',
  USERS = 'USERS',
  SETTINGS = 'SETTINGS',
  LICENSE_GEN = 'LICENSE_GEN',
  TIME_CLOCK = 'TIME_CLOCK'
}

export type UserRole = 'ADMIN' | 'MANAGER' | 'ACCOUNTANT' | 'SALES' | 'HR_MANAGER';
export type NetworkRole = 'HUB' | 'WORKSTATION' | 'STANDALONE';
export type LicenseType = 'STARTUP' | 'RETAIL' | 'SERVICE' | 'HR_FOCUS' | 'EVENTS' | 'BUSINESS' | 'CORPORATE';

export interface Traceable {
  createdBy?: string;
  createdById?: string;
  updatedAt?: number;
  updatedBy?: string;
  isDeleted?: boolean;
}

export interface AuditLog {
  id: string;
  timestamp: number;
  userId: string;
  userName: string;
  action: 'CREATE' | 'UPDATE' | 'DELETE' | 'SYNC';
  module: ModuleType;
  details: string;
}

export interface User extends Traceable {
  id: string;
  username: string;
  name: string;
  role: UserRole;
  email: string;
  password?: string;
  active: boolean;
}
// ...
export interface Product extends Traceable {
  id: string;
  code: string;
  name: string;
  priceRetail: number;
  priceWholesale?: number;
  rentalPriceDay: number;
  stock: number;
  minStock: number;
  unit: string;
  category: string;
  isRental: boolean;
  isSale: boolean;
  isPhysical: boolean;
  status: 'Available' | 'Rented' | 'Maintenance' | 'Broken';
  expiryDate?: string;
  lotNumber?: string;
}